Bug 20133: Enable/disable the "Hide patron information" feature per group of libraries
In order to control this feature correctly it needs to be enabled for
group of libraries and not for all groups defined in the system.
Groups will be used for different usages and so the feature must not be
enabled by default for all of them.

Test plan:
Retest bug 18403 with the feature turned on/off for a given tree

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>