Bug 20345: Put saved report keyword search form on reports home
This patch makes cosmetic changes to the reports home page in the staff
client following the pattern set on the Administration home page with
the system preferences search box. A reports keyword search form is
added, and reports actions have font-awesome icons.

To test, apply the patch and view the reports home page. Confirm that
all affected links work correctly and that the search form works
correctly.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>