Bug 20268: (follow-up) Correct template error
The login template must have a class attribute on the body tag in order
for the template to pass tests.

To test, apply the patch and confirm that the staff client login form
still looks the same. For further confirmation you could update the
IntranetUserCSS system preference with something like this:

.main_main-auth {
    background-color: #CCF;
}

The login form should now have a different background color.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>