Bug 20659: Fix blocking errors display
The include file blocking-errors.inc has been created to by-pass regular
processing and display an error that should block the elements of the
view.

For instance you call a patron's page with invalid borrowernumber => we
do not want the app to crash (500) or the page to display with broken
elements.

It worked well on bug 18403 but it's now broken, a blank page is
displayed instead.
It may be caused by the move of the JS to the bottom of the page.

Test plan:
hit /cgi-bin/koha/members/moremember.pl?borrowernumber=45432
Without this patch a blank page is displayed.
With this patch you see the blocking error

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>